style: Rename package name
