--- core/src/main/java/com/tlcsdm/core/modal/VoskModal.java
@@ -1,20 +1,0 @@
-package com.tlcsdm.core.modal;
-
-import java.io.Serializable;
-
-/**
- * vosk 结果.
- *
- * @author unknowIfGuestInDream
- */
-public class VoskModal implements Serializable {
-    private String text;
-
-    public String getText() {
-        return text;
-    }
-
-    public void setText(String text) {
-        this.text = text;
-    }
-}
